+++ shapeless/src/main/scala/better/files/ShapelessScanner.scala
@@ -9,6 +9,6 @@ object ShapelessScanner {
   implicit def hListScannable[H, T <: HList](implicit h: Scannable[H], t: Scannable[T]): Scannable[H :: T] =
     Scannable(s => h(s) :: t(s))
 
-  implicit def caseClassParser[A, R](implicit gen: Generic.Aux[A, R], reprScannable: Scannable[R]): Scannable[A] =
+  implicit def genericScannable[A, R](implicit gen: Generic.Aux[A, R], reprScannable: Scannable[R]): Scannable[A] =
     Scannable(s => gen.from(reprScannable(s)))
 }